+++ Classes/GroupMgt.py
@@ -8,6 +8,7 @@
 ListOfGroups[group id]['Name']    - Group Name as it will be created in Domoticz
 ListOfGroups[group id]['Devices'] - List of Devices associed to this group on Zigate
 ListOfGroups[group id]['Imported']- List of Devices to be associated to the group. We might have some removal, or some addiional from previous run
+ListOfGroups[group id]['Tradfri Remote']- Manage the Tradfri Remote
 
 self.ListOfDevices[nwkid]['GroupMgt'][Ep][GroupID]['Phase'] = 'OK-Membership' / 'REQ-Membership' / 'DEL-Membership'
 self.ListOfDevices[nwkid]['GroupMgt'][Ep][GroupID]['Phase-Stamp'] = time()
@@ -179,8 +180,23 @@ class GroupsManagement(object):
                             # Check if this is not the Zigate itself
                             Domoticz.Error("load_ZigateGroupConfiguration - Unknown address %s to be imported" %_ieee )
                             continue
-                        # Let's check if we don't have the EP included as well
-                        self.ListOfGroups[group_id]['Imported'].append( (_ieee, _ieeeEp) )
+                        # Finaly, let's check if this is not an IKEA Tradfri Remote
+                        nwkid = self.IEEE2NWK[_ieee]
+                        _tradfri_remote = False
+                        if 'Ep' in self.ListOfDevices[nwkid]:
+                            if '01' in self.ListOfDevices[nwkid]['Ep']:
+                                if 'ClusterType' in self.ListOfDevices[nwkid]['Ep']['01']:
+                                    for iterDev in self.ListOfDevices[nwkid]['Ep']['01']['ClusterType']:
+                                        if self.ListOfDevices[nwkid]['Ep']['01']['ClusterType'][iterDev] == 'Ikea_Round_5b':
+                                            # We should not process it through the group.
+                                            Domoticz.Log("Not processing Ikea Tradfri as part of Group. Will enable the Left/Right actions")
+                                            self.ListOfGroups[group_id]['Tradfri Remote'] = {}
+                                            self.ListOfGroups[group_id]['Tradfri Remote']['Device Addr'] = nwkid
+                                            self.ListOfGroups[group_id]['Tradfri Remote']['Device Id'] = iterDev
+                                            _tradfri_remote = True
+                        if not _tradfri_remote:
+                            # Let's check if we don't have the EP included as well
+                            self.ListOfGroups[group_id]['Imported'].append( (_ieee, _ieeeEp) )
 
                     Domoticz.Debug(" )> Group Imported: %s" %group_name)
             if group_id :
@@ -520,6 +536,7 @@ class GroupsManagement(object):
                 'ColorControlFull':5 }    # ( 241, 7, 7) - Like RGBWW, but allows combining RGB and white
 
         code = 0
+        _ikea_colormode = None
         color_widget = None
         widget = ( 241, 7,7 )
         for devNwkid, devEp in self.ListOfGroups[group_nwkid]['Devices']:
@@ -531,12 +548,18 @@ class GroupsManagement(object):
                     devwidget = self.ListOfDevices[devNwkid]['Ep'][devEp]['ClusterType'][iterClusterType]
                     if code <= WIDGETS[devwidget]:
                         code = WIDGETS[devwidget]
-                        if code == 1: widget = ( 244, 73, 0 )
-                        elif code == 2: widget = ( 244, 73, 7 )
+                        if code == 1: 
+                            widget = ( 244, 73, 0 )
+                        elif code == 2: 
+                            widget = ( 244, 73, 7 )
                         elif code == 3 :
                             if color_widget is None:
-                                if devwidget == 'ColorControlWW': widget = ( 241, 8, 7 )
-                                elif devwidget == 'ColorControlRGB': widget = ( 241, 2, 7 )
+                                if devwidget == 'ColorControlWW': 
+                                    widget = ( 241, 8, 7 )
+                                    _ikea_colormode = devwidget
+                                elif devwidget == 'ColorControlRGB': 
+                                    widget = ( 241, 2, 7 )
+                                    _ikea_colormode = devwidget
                             elif color_widget == devwidget:
                                 continue
                             elif (devwidget == 'ColorControlWW' and color_widget == 'ColorControlRGB') or \
@@ -544,9 +567,15 @@ class GroupsManagement(object):
                                 code = 4
                                 color_widget = 'ColorControlRGBWW'
                                 widget = ( 241, 4, 7)
+                                _ikea_colormode = color_widget
                         elif code == 4: widget = ( 241, 4, 7)
                         elif code == 5: widget = ( 241, 7, 7)
                     pre_code = code
+
+
+        # This will be used when receiving left/right click , to know if it is RGB or WW
+        if 'Tradfri Remote' in self.ListOfGroups[group_nwkid]:
+            self.ListOfGroups[group_nwkid]['Tradfri Remote']['Color Mode'] = _ikea_colormode
 
         Domoticz.Debug("_bestGroupWidget - Code: %s, Color_Widget: %s, widget: %s" %( code, color_widget, widget))
         return widget
@@ -653,6 +682,42 @@ class GroupsManagement(object):
 
         return
 
+    def set_Kelvin_Color( self, mode, addr, EPin, EPout, t, transit=None):
+        #Value is in mireds (not kelvin)
+        #Correct values are from 153 (6500K) up to 588 (1700K)
+        # t is 0 > 255
+    
+        if transit is None:
+            transit = '0001'
+        else:
+            transit = '%04x' %transit
+
+        TempKelvin = int(((255 - int(t))*(6500-1700)/255)+1700)
+        TempMired = 1000000 // TempKelvin
+        zigate_cmd = "00C0"
+        zigate_param = Hex_Format(4,TempMired) + transit
+        datas = "%02d" %mode + addr + EPin + EPout + zigate_param
+        Domoticz.Debug("Command: %s - data: %s" %(zigate_cmd,datas))
+        self.ZigateComm.sendData( zigate_cmd, datas)
+
+    def set_RGB_color( self, mode, addr, EPin, EPout, r, g, b, transit=None):
+
+        if transit is None:
+            transit = '0001'
+        else:
+            transit = '%04x' %transit
+        x, y = rgb_to_xy((int(r),int(g),int(b)))
+        #Convert 0>1 to 0>FFFF
+        x = int(x*65536)
+        y = int(y*65536)
+        strxy = Hex_Format(4,x) + Hex_Format(4,y)
+        zigate_cmd = "00B7"
+        zigate_param = strxy + transit
+        datas = "%02d" %mode + addr + EPin + EPout + zigate_param
+        Domoticz.Debug("Command: %s - data: %s" %(zigate_cmd,datas))
+        self.ZigateComm.sendData( zigate_cmd, datas)
+
+
     def processCommand( self, unit, nwkid, Command, Level, Color_ ) : 
 
         Domoticz.Debug("processCommand - unit: %s, nwkid: %s, cmd: %s, level: %s, color: %s" %(unit, nwkid, Command, Level, Color_))
@@ -724,29 +789,13 @@ class GroupsManagement(object):
                 Domoticz.Debug("Not implemented device color 1")
             #ColorModeTemp = 2   // White with color temperature. Valid fields: t
             if Hue_List['m'] == 2:
-                #Value is in mireds (not kelvin)
-                #Correct values are from 153 (6500K) up to 588 (1700K)
-                # t is 0 > 255
-                TempKelvin = int(((255 - int(Hue_List['t']))*(6500-1700)/255)+1700);
-                TempMired = 1000000 // TempKelvin
-                zigate_cmd = "00C0"
-                zigate_param = Hex_Format(4,TempMired) + "0000"
-                datas = "%02d" %ADDRESS_MODE['group'] + nwkid + EPin + EPout + zigate_param
-                Domoticz.Debug("Command: %s - data: %s" %(zigate_cmd,datas))
-                self.ZigateComm.sendData( zigate_cmd, datas)
+                self.set_Kelvin_Color( ADDRESS_MODE['group'], nwkid, EPin, EPout, int(Hue_List['t']))
 
             #ColorModeRGB = 3    // Color. Valid fields: r, g, b.
             elif Hue_List['m'] == 3:
-                x, y = rgb_to_xy((int(Hue_List['r']),int(Hue_List['g']),int(Hue_List['b'])))
-                #Convert 0>1 to 0>FFFF
-                x = int(x*65536)
-                y = int(y*65536)
-                strxy = Hex_Format(4,x) + Hex_Format(4,y)
-                zigate_cmd = "00B7"
-                zigate_param = strxy + "0000"
-                datas = "%02d" %ADDRESS_MODE['group'] + nwkid + EPin + EPout + zigate_param
-                Domoticz.Debug("Command: %s - data: %s" %(zigate_cmd,datas))
-                self.ZigateComm.sendData( zigate_cmd, datas)
+
+                self.set_RGB_color( ADDRESS_MODE['group'], nwkid, EPin, EPout, \
+                        int(Hue_List['r']), int(Hue_List['g']), int(Hue_List['b']))
 
             #ColorModeCustom = 4, // Custom (color + white). Valid fields: r, g, b, cw, ww, depending on device capabilities
             elif Hue_List['m'] == 4:
@@ -782,6 +831,71 @@ class GroupsManagement(object):
                 sValue = str(value)
                 self.Devices[unit].Update(nValue=int(nValue), sValue=str(sValue), Color=Color_) 
                 return
+
+    def manageIkeaTradfriRemoteLeftRight( self, addr, type_dir):
+
+        for iterGrp in self.ListOfGroups:
+            if 'Tradfri Remote' not in self.ListOfGroups[iterGrp]:
+                continue
+            if addr != self.ListOfGroups[iterGrp]['Tradfri Remote']['Device Addr']:
+                continue
+            _grpid = iterGrp
+            break
+        else:
+            _ieee = self.ListOfDevices[addr]['IEEE']
+            Domoticz.Log("manageIkeaTradfriRemoteLeftRight - Remote %s not associated to any group" %_ieee)
+            return
+            
+        _widgetColor = self.ListOfGroups[_grpid]['Tradfri Remote']['Color Mode'] 
+        Domoticz.Log("manageIkeaTradfriRemoteLeftRight - Color model : %s" %_widgetColor)
+
+        if _widgetColor in ('ColorControlWW'): # Will work in Kelvin
+            if 'Actual T' not in self.ListOfGroups[_grpid]['Tradfri Remote']:
+                t = 128
+            else:
+                t = self.ListOfGroups[_grpid]['Tradfri Remote']['Actual T']
+
+            if type_dir == 'left':
+                t -= self.pluginconf.TradfriKelvinStep
+                if t < 0: t = 255
+            elif type_dir == 'right':
+                t += self.pluginconf.TradfriKelvinStep
+                if t > 255: t = 0
+                
+            Domoticz.Log("manageIkeaTradfriRemoteLeftRight - Kelvin T %s" %t)
+            self.set_Kelvin_Color( ADDRESS_MODE['group'], _grpid, '01', '01', t)
+            self.ListOfGroups[_grpid]['Tradfri Remote']['Actual T'] = t
+
+        elif _widgetColor in ('ColorControlRGB','ColorControlRGBWW', 'ColorControl', 'ColorControlFull'): # Work in RGB
+            # Here we will scroll R, G and B 
+
+            PRESET_COLOR = (  
+                              (  10,  10,  10), # 
+                              ( 255,   0,   0), # Red
+                              (   0, 255,   0), # Green
+                              (   0,   0, 255), # Blue
+                              ( 255, 255,   0), # Yello
+                              (   0, 255, 255), # Aqua
+                              ( 255,   0, 255), # 
+                              ( 255, 255, 255)  # Whhite
+                           )
+
+            if 'RGB' not in self.ListOfGroups[_grpid]['Tradfri Remote']:
+                seq_idx = 0
+            else:
+                seq_idx = self.ListOfGroups[_grpid]['Tradfri Remote']['RGB']
+
+            r, g, b = PRESET_COLOR[seq_idx]
+
+            if type_dir == 'left': seq_idx -= 1
+            elif type_dir == 'right': seq_idx += 1
+
+            if seq_idx >= len(PRESET_COLOR): seq_idx = 0
+            if seq_idx < 0: seq_idx = len(PRESET_COLOR) - 1
+
+            Domoticz.Log("manageIkeaTradfriRemoteLeftRight - R %s G %s B %s" %(r,g,b))
+            self.set_RGB_color( ADDRESS_MODE['group'], _grpid, '01', '01', r, g, b)
+            self.ListOfGroups[_grpid]['Tradfri Remote']['RGB'] = seq_idx
 
     def hearbeatGroupMgt( self ):
         ' hearbeat to process Group Management actions '
@@ -1154,6 +1268,7 @@ class GroupsManagement(object):
                             self.ListOfGroups[iterGrp]['Name'] = self.Devices[x].Name
                             # Check if we need to update the Widget
                             self._updateDomoGroupDeviceWidget(self.ListOfGroups[iterGrp]['Name'], iterGrp)
+                            Domoticz.Log("hearbeatGroupMgt - _updateDomoGroup done")
                         break
                 else:
                     # Unknown group in Domoticz. Create it
@@ -1164,7 +1279,9 @@ class GroupsManagement(object):
                         self.ListOfGroups[iterGrp]['Name'] = "Zigate Group %s" %iterGrp
                     Domoticz.Log("hearbeatGroupMgt - create Domotciz Widget for %s " %self.ListOfGroups[iterGrp]['Name'])
                     self._createDomoGroupDevice( self.ListOfGroups[iterGrp]['Name'], iterGrp)
+
             self.StartupPhase = 'end of group startup'
+            self._write_GroupList()
 
         elif self.StartupPhase == 'end of group startup':
             for iterGrp in self.ListOfGroups: