+++ Modules/output.py
@@ -27,12 +27,12 @@ def ZigatePermitToJoin( self, permit ):
 
     if permit:
         self.permitTojoin = permit
-        Domoticz.Log("Zigate enter in discovery mode for %s seconds " %permit)
+        Domoticz.Log("Request discovery mode for %s seconds " %permit)
         sendZigateCmd(self, "0049","FFFC" + '%02x' %permit + "00")
         sendZigateCmd( self, "0014", "" ) # Request status
     else: 
         self.permitTojoin = 0x00
-        Domoticz.Log("Zigate stop discovery mode")
+        Domoticz.Log("Request stop discovery mode")
         #sendZigateCmd(self, "0049","FFFC" + '01' + "00")
         sendZigateCmd( self, "0014", "" ) # Request status
 
@@ -261,9 +261,9 @@ def ReadAttributeRequest_0000(self, key, fullScope=True):
             Domoticz.Debug("Request Basic  via Read Attribute request %s/%s %s" %(key, EPout, str(listAttributes)))
             ReadAttributeReq( self, key, EPin, EPout, "0000", listAttributes )
         else:
-            Domoticz.Log("Request Basic  via Read Attribute request part1 %s/%s %s" %(key, EPout, str(listAttr1)))
+            Domoticz.Debug("Request Basic  via Read Attribute request part1 %s/%s %s" %(key, EPout, str(listAttr1)))
             ReadAttributeReq( self, key, EPin, EPout, "0000", listAttr1 )
-            Domoticz.Log("Request Basic  via Read Attribute request part2 %s/%s %s" %(key, EPout, str(listAttr2)))
+            Domoticz.Debug("Request Basic  via Read Attribute request part2 %s/%s %s" %(key, EPout, str(listAttr2)))
             ReadAttributeReq( self, key, EPin, EPout, "0000", listAttr2 )
 
 
@@ -399,7 +399,7 @@ def ReadAttributeRequest_0102(self, key):
 
     EPin = "01"
     EPout= "01"
-    Domoticz.Log("Request Windows Covering status Read Attribute request: " + key )
+    Domoticz.Debug("Request Windows Covering status Read Attribute request: " + key )
     listAttributes = []
 
     listAttributes.append(0x0000) # Window Covering Type
@@ -573,7 +573,7 @@ def removeZigateDevice( self, IEEE ):
         Domoticz.Status("Remove from Zigate Device = " + " IEEE = " +str(IEEE) )
         #sendZigateCmd(self, "0026", str(self.ZigateIEEE) + str(IEEE) )
     else:
-        Domoticz.Log("removeZigateDevice - cannot remove due to unknown Zigate IEEE: ")
+        Domoticz.Error("removeZigateDevice - cannot remove due to unknown Zigate IEEE: ")
 
     return
 
@@ -614,7 +614,7 @@ def processConfigureReporting( self, NWKID=None ):
                                  '0021': {'DataType': '29', 'MinInterval':'0E10', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'}}},
 
         # On/Off Cluster
-        '0006': {'Attributes': { '0000': {'DataType': '10', 'MinInterval':'0001', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}}},
+        '0006': {'Attributes': { '0000': {'DataType': '10', 'MinInterval':'0005', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}}},
         #'0006': {'Attributes': { '0000': {'DataType': '10', 'MinInterval':'0003', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'00'}}},
 
         # Level Control Cluster
@@ -642,12 +642,12 @@ def processConfigureReporting( self, NWKID=None ):
                                  '001B': {'DataType': '30', 'MinInterval':'012C', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'},
                                  '001C': {'DataType': '30', 'MinInterval':'012C', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'}}},
         # Colour Control
-        '0300': {'Attributes': { '0007': {'DataType': '21', 'MinInterval':'0384', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'},
-                                 '0000': {'DataType': '20', 'MinInterval':'0384', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'},
-                                 '0001': {'DataType': '20', 'MinInterval':'0384', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'},
-                                 '0003': {'DataType': '21', 'MinInterval':'0384', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'},
-                                 '0004': {'DataType': '21', 'MinInterval':'0384', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'},
-                                 '0008': {'DataType': '30', 'MinInterval':'0384', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'}}},
+        '0300': {'Attributes': { '0007': {'DataType': '21', 'MinInterval':'0001', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}, # Color Temp
+                                 #'0000': {'DataType': '20', 'MinInterval':'0384', 'MaxInterval':'0E10', 'TimeOut':'0FFF','Change':'01'},
+                                 #'0001': {'DataType': '20', 'MinInterval':'0001', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}, 
+                                 '0003': {'DataType': '21', 'MinInterval':'0001', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}, # Color X
+                                 '0004': {'DataType': '21', 'MinInterval':'0001', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}, # Color Y
+                                 '0008': {'DataType': '30', 'MinInterval':'0001', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}}}, # Color Mode
         # Illuminance Measurement
         '0400': {'Attributes': { '0000': {'DataType': '21', 'MinInterval':'0005', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'0F'}}},
         # Temperature
@@ -670,7 +670,7 @@ def processConfigureReporting( self, NWKID=None ):
                                  '0002': {'DataType': '19', 'MinInterval':'003C', 'MaxInterval':'0384', 'TimeOut':'0FFF','Change':'01'}}},
         # Power
         '0702': {'Attributes': { '0000': {'DataType': '25', 'MinInterval':'FFFF', 'MaxInterval':'0000', 'TimeOut':'0000','Change':'00'},
-                                 '0400': {'DataType': '2a', 'MinInterval':'003C', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}}}
+                                 '0400': {'DataType': '2a', 'MinInterval':'0005', 'MaxInterval':'012C', 'TimeOut':'0FFF','Change':'01'}}}
         }
 
     now = int(time())
@@ -926,30 +926,30 @@ def identifySend( self, nwkid, ep, duration=0):
     sendZigateCmd(self, "0070", datas )
 
 def maskChannel( channel ):
+
     CHANNELS = { 0: 0x00000000, # Scan for all channels
             11: 0x00000800,
-            12: 0x00001000,
-            13: 0x00002000,
-            14: 0x00004000,
+            #12: 0x00001000,
+            #13: 0x00002000,
+            #14: 0x00004000,
             15: 0x00008000,
-            16: 0x00010000,
-            17: 0x00020000,
-            18: 0x00040000,
+            #16: 0x00010000,
+            #17: 0x00020000,
+            #18: 0x00040000,
             19: 0x00080000,
             20: 0x00100000,
-            21: 0x00200000,
-            22: 0x00400000,
-            23: 0x00800000,
-            24: 0x01000000,
+            #21: 0x00200000,
+            #22: 0x00400000,
+            #23: 0x00800000,
+            #24: 0x01000000,
             25: 0x02000000,
             26: 0x04000000 }
 
     mask = 0x00000000
-    Domoticz.Debug("setChannel - Channel list: %s" %(channel))
     if isinstance(channel, list):
         for c in channel:
             if c.isdigit():
-                if c in ( '0', '11','12','13','14','15','16','17','18','19','20','21','22','23','24','25','26'):
+                if int(c) in CHANNELS:
                     mask += CHANNELS[int(c)]
             else:
                 Domoticz.Error("maskChannel - invalid channel %s" %c)
@@ -1049,17 +1049,20 @@ def setExtendedPANID(self, extPANID):
             %( extPANID) )
     sendZigateCmd(self, "0020", datas )
 
-def leaveMgtReJoin( self, saddr, ieee):
+def leaveMgtReJoin( self, saddr, ieee, rejoin=True):
     ' in case of receiving a leave, and that is not related to an explicit remove '
-
-    if self.permitTojoin != 0xff:
-        Domoticz.Log("Switch to Permit to Join for 30s")
-        discover = "%02.X" %int(30)
-        sendZigateCmd(self, "0049","FFFC" + discover + "00")
 
     Domoticz.Log("leaveMgt - sAddr: %s , ieee: %s" %( saddr, ieee))
     # Request a Re-Join and Do not remove children
-    datas = saddr + ieee + '01' + '00'
+    if rejoin:
+        if self.permitTojoin != 0xff:
+            Domoticz.Log("Switch to Permit to Join for 30s, to allow rejoin")
+            discover = "%02.X" %int(30)
+            sendZigateCmd(self, "0049","FFFC" + discover + "00")
+        datas = saddr + ieee + '01' + '00'
+    else:
+        datas = saddr + ieee + '00' + '00'
+
     sendZigateCmd(self, "0047", datas )
 
 def thermostat_Setpoint_SPZB(  self, key, setpoint):
@@ -1183,7 +1186,7 @@ def thermostat_Mode( self, key, mode ):
 
 def ReadAttributeRequest_0201(self, key):
 
-    Domoticz.Log("ReadAttributeRequest_0201 - Key: %s " %key)
+    Domoticz.Debug("ReadAttributeRequest_0201 - Key: %s " %key)
     # Power Config
     EPin = "01"
     EPout= "01"
@@ -1209,7 +1212,7 @@ def ReadAttributeRequest_0201(self, key):
 
     listAttributes = []
     if str(self.ListOfDevices[key]['Model']).find('SPZB') == 0:
-        Domoticz.Log("- req Attributes for Eurotronic")
+        Domoticz.Debug("- req Attributes for Eurotronic")
         listAttributes.append(0x4000)        # TRV Mode
         listAttributes.append(0x4001)        # Set Valve Position
         listAttributes.append(0x4002)        # Errors